move components into dir
