+++ src/app/users/components/sign-in/sign-in.component.scss
@@ -1,4 +1,4 @@
-@import '~@angular/material/theming';
+@import '../../../../../node_modules/@angular/material/theming';
 
 :host {
     .flex-container {
