+++ src/app/users/users-routing.module.ts
@@ -2,10 +2,10 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
 // components
-import { MyAccountComponent } from './my-account/my-account.component';
-import { SignInComponent } from './sign-in/sign-in.component';
-import { SignOutComponent } from './sign-out/sign-out.component';
-import { SignUpComponent } from './sign-up/sign-up.component';
+import { MyAccountComponent } from './components/my-account/my-account.component';
+import { SignInComponent } from './components/sign-in/sign-in.component';
+import { SignOutComponent } from './components/sign-out/sign-out.component';
+import { SignUpComponent } from './components/sign-up/sign-up.component';
 
 import { AuthenticatedGuard } from '../shared/authentication.guard';
